DHCP timer now defaults to the system timer if not specified

+++ src/uip/dhcpc.h
@@ -40,7 +40,7 @@ struct dhcpc_state {
   struct pt pt;
   char state;
   struct uip_udp_conn *conn;
-  u32 timer_init, ticks;
+  u64 timer_init, ticks;
   const void *mac_addr;
   int mac_len;
